Update location fields in StallDto, StallLocationUpdateRequest, and Stall to use Double instead of Integer

+++ stall-service/src/main/java/com/bookfair/stallservice/dto/StallDto.java
@@ -23,6 +23,6 @@ public class StallDto {
 
     private Boolean isReserved = false;
 
-    private Integer locationX;
-    private Integer locationY;
+    private Double locationX;
+    private Double locationY;
 }
